Clean up the code style
This is a first pass at cleaning up all the indentation and some code
style fixes. Indentation format was standardized on tabs as they take
less characters on the files that are sent out to clients (and are
adjustable in editors).

There is quite a bit of duplicate code, subsequent commits will attempt
to clean that up.