Move to a single get-info-modal
Modifies serverdatapdo.php to send the table passed to it to the modal
source (and fixes a few errors about undefined variables that were
spamming the log on every call).
Makes get-info-modal.php get its table from the ajax call and sets
user.php to use the unified data source.