Consolidate recently_added.php
Note, there is a slight change in behavior for this. If the content type
is movie it used to set the <a href> to $recentXml-Video['ratingKey'] if
a thumbnail wasn't able to be found for the item. It now no matter what
uses $recentXml['ratingKey']. Current PMS versions seem to behave this
way in either case, and it seem the other behavior may have been a
copy/paste relic.