Send errors back to the settings page
When the process_settings.php page runs into a problem with the posted
values it now sends the error back to the settings page (and the PHP
error log).
Also causes the settings page to _not_ display the welcome modal when an
error was present in the $_GET, even when no current settings file
exists.