Grab all data from PMS using cURL
file_get_contents() and simplexml_load_string() don't properly support
HTTPS, so move to a single function to gather data from a PMS server.
This allows some simplification as in most cases figuring out the proper
PMS URL and token needs to only be done in this single place now.