Re-implement settings
Move to using a JSON file for the settings, and do all relations with
the settings through a class instead of a globally included variable.
Test the settings before attempting to use them.

This is the first part of implementation, and only includes the basics.
More tests will need to be implemented, specifically validation of the
$_POST data.