Check if the settings file exists on the index
Since the index.php file never actually directly uses the settings, but
calls files that do require it but can't properly redirect the user as
they are called via AJAX, it needs to at least check that the settings
file is there before doing the rest of the page.

Also works around a bug that triggers on PHP < 5.4.0.