Handle Token de-authentication
Somewhat gracefully handle token de-authorization, redirecting the user
to the settings page with a message about authorization failure.

This also revamps the settings to _never_ store the user/password. Since
we don't need them ever again after getting an authentication token, and
if the token is invalid we would need them specified again it is better
to just have the user enter them again then take on the responsibility
of storing them properly.