Guard against writing a bad config
Create a flag that, when set, prevents writing the settings file.
Previously on the very first save of the file it was possible to write
the config file with invalid data.

Ensures that when a token is not yet set (or blank) it will attempt to
set a new one.

Also adds a debug dumping function for testing, with output disabled by
default.